Extract type definitions from dataloader
